+++ src/lib.test.ts
@@ -20,10 +20,43 @@ import { getRxStoragePESQLite } from "./lib";
 
 describe("pe-sqlite-for-rxdb tests", () => {
   it("quickstart works", async () => {
+    // Create the RxDatabase
     const myDatabase = await createRxDatabase({
-      const: "myDatabase",
+      name: "myDatabase",
       storage: getRxStoragePESQLite(),
     });
     expect(myDatabase).toBeTruthy();
+
+    // Create an RxCollection
+    // Creating a schema for a collection
+    const todoSchema = {
+      version: 0,
+      primaryKey: "id",
+      type: "object",
+      properties: {
+        id: {
+          type: "string",
+          maxLength: 100,
+        },
+        name: {
+          type: "string",
+        },
+        done: {
+          type: "boolean",
+        },
+        timestamp: {
+          type: "string",
+          format: "date-time",
+        },
+      },
+      required: ["id", "name", "done", "timestamp"],
+    };
+
+    // Adding an RxCollection to the RxDatabase
+    await myDatabase.addCollections({
+      todos: {
+        schema: todoSchema,
+      },
+    });
   });
 });